feat: Refactor More section to use dedicated screens for About Us, Contact Info, FAQ, and Navigation for Teams
- Added new screens for About Us, Contact Info, FAQ, and Navigation for Teams.
- Removed modal components for About Us, Contact Info, FAQ, and Navigation for Teams.
- Updated MoreScreen to navigate to the new screens instead of using modals.
- Enhanced styling and structure for the new screens.
- Implemented status bar management in AnnualReportModal.
